devops-projects/techshop-v1/services/user-service/requirements.txt
 Flask
 Flask-SQLAlchemy
 PyJWT
 bcrypt
 marshmallow
 psycopg2-binary
 python-dotenv